% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamstaffel.R
\name{get_results}
\alias{get_results}
\title{Get Teamstaffel Results Table}
\usage{
get_results(
  sex = "M",
  year = 2022,
  event = "ALL",
  base_url = "https://berlin-wasserbetriebe.r.mikatiming.de"
)
}
\arguments{
\item{sex}{either "M" for men or "W" for women default: "M"}

\item{year}{default: 2022}

\item{event}{return overall results ("ALL") or of a selected starting day (i.e.
"ST1", "ST2" or "ST3"), (default: "ALL")}

\item{base_url}{default: "https://berlin-wasserbetriebe.r.mikatiming.de"}
}
\value{
tibble with results
}
\description{
Get Teamstaffel Results Table
}
\examples{
women <- get_results("W")
women
}
